chore(refactor): turned a few "unique" components from widgets to "root" components
